+++ src/v1.x/packages/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @copilotkit/a2ui-renderer
 
+## 1.51.0-next.4
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.0-next.4
+
+## 1.51.0-next.3
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.0-next.3
+
+## 1.51.0-next.2
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.0-next.2
+
+## 1.51.0-next.1
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.0-next.1
+
 ## 1.50.2-next.0
 
 ### Patch Changes